Scripts merged, backup is done from a copy and everything is done from a single directory
