+++ .idea/workspace.xml
@@ -3,9 +3,10 @@
   <component name="ChangeListManager">
     <list default="true" id="848c41f3-5df6-4af5-91df-99649f63470f" name="Changes" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/out/production/createVacuumScript/Vacuum.class" beforeDir="false" afterPath="$PROJECT_DIR$/out/production/createVacuumScript/Vacuum.class" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/out/production/createVacuumScript/Backup.class" beforeDir="false" afterPath="$PROJECT_DIR$/out/production/createVacuumScript/Backup.class" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/out/production/createVacuumScript/Vacuum.class" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/src/Backup.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/Backup.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/Vacuum.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/Vacuum.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/Vacuum.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -45,7 +46,7 @@
     <property name="project.structure.side.proportion" value="0.0" />
     <property name="settings.editor.selected.configurable" value="preferences.pluginManager" />
   </component>
-  <component name="RunManager" selected="Application.Backup">
+  <component name="RunManager" selected="Application.main">
     <configuration name="Backup" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
       <option name="MAIN_CLASS_NAME" value="Backup" />
       <module name="createVacuumScript" />
@@ -53,8 +54,8 @@
         <option name="Make" enabled="true" />
       </method>
     </configuration>
-    <configuration name="Vacuum" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
-      <option name="MAIN_CLASS_NAME" value="Vacuum" />
+    <configuration name="main" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
+      <option name="MAIN_CLASS_NAME" value="main" />
       <module name="createVacuumScript" />
       <method v="2">
         <option name="Make" enabled="true" />
@@ -62,8 +63,8 @@
     </configuration>
     <recent_temporary>
       <list>
+        <item itemvalue="Application.main" />
         <item itemvalue="Application.Backup" />
-        <item itemvalue="Application.Vacuum" />
       </list>
     </recent_temporary>
   </component>
